Removed ChannelService, more fixes
