+++ util/src/index.ts
@@ -4,7 +4,6 @@ import "reflect-metadata";
 export * from "./util/index";
 export * from "./interfaces/index";
 export * from "./entities/index";
-export * from "./services/index";
 export * from "./dtos/index";
 
 // import Config from "../util/Config";
