--- util/src/services/index.ts
@@ -1,1 +1,0 @@
-export * from "./ChannelService";
